requirements.txt
fastapi>=0.115,<1
uvicorn[standard]>=0.30,<1
gradio>=5.0,<6
python-multipart>=0.0.9,<1
pillow>=10,<12
numpy>=1.26,<3
pandas>=2.2,<3
scikit-learn>=1.4,<2
pydantic>=2.7,<3
tensorflow>=2.16,<2.18
